playing around inventory

+++ DioriteAPI/src/main/java/org/diorite/material/blocks/earth/DirtMat.java
@@ -136,7 +136,7 @@ public class DirtMat extends EarthMat implements VariantableMat
      */
     public static void register(final DirtMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((byte) element.getType(), element);
         byName.put(element.name(), element);
     }
