+++ DioriteAPI/src/main/java/org/diorite/material/blocks/redstone/GoldenPressurePlateMat.java
@@ -148,7 +148,7 @@ public class GoldenPressurePlateMat extends WeightedPressurePlateMat
      */
     public static void register(final GoldenPressurePlateMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((byte) element.getType(), element);
         byName.put(element.name(), element);
     }
