+++ DioriteAPI/src/main/java/org/diorite/material/blocks/redstone/piston/PistonHeadMat.java
@@ -157,7 +157,7 @@ public class PistonHeadMat extends PistonBaseMat
      */
     public static void register(final PistonHeadMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((byte) element.getType(), element);
         byName.put(element.name(), element);
     }
