+++ DioriteAPI/src/main/java/org/diorite/material/blocks/stony/ore/CoalOreMat.java
@@ -101,7 +101,7 @@ public class CoalOreMat extends OreMat
      */
     public static void register(final CoalOreMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((byte) element.getType(), element);
         byName.put(element.name(), element);
     }
